requirements.txt
Django>=5.2,<5.3
wagtail>=7.0,<7.1
django-cloudinary-storage==0.3.0
django-colorfield==0.14.0
django-crispy-forms==2.4
django-filter==25.1
django-localflavor==4.0
django-modelcluster==6.4
django-parler==2.3
channels==4.2.2
django-permissionedforms==0.1
weasyprint==61.2
dj-database-url==2.3.0 
redis==5.0.4
django-redis==5.4.0
django-rosetta==0.10.0
django-stubs-ext==5.2.0
django-taggit==6.1.0
django-tasks==0.7.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
crispy-bootstrap5
django-import-export
